requirements.txt
torch
torchvision
tqdm
scipy 
sklearn 
numpy
PyYAML

